requirements.txt
Django>=4.0
djangorestframework
psycopg2-binary
python-decouple
djangorestframework-simplejwt
pytest
pytest-django
pytest-cov
